Move `SerializeTable` to `Client.Common`, add unit test
